REFACTOR: moved the user dashboard folder into the components folder
